refactor(skill-forge): split runner tests and relocate temp claude homes
tests/test_run_eval.py had grown past 1,100 lines across three distinct
concerns; it is now split into aggregation/dispatch/masking tests
(test_run_eval.py), the Claude runner (test_run_eval_claude.py), and the
Codex runner (test_run_eval_codex.py), each importing its subject module
directly.

Temp claude homes also move from the project root to system temp so a
hard-killed eval cannot leave skill-forge-claude-home-* directories in
the user's repository; the location assertions live in the split files,
which is why both changes land together.
