+++ plugins/agent-skills/skills/skill-forge/tests/test_run_eval.py
@@ -1,24 +1,11 @@
-"""Tests for scripts.run_eval module."""
+"""Tests for scripts.run_eval aggregation, dispatch, and masking."""
 
-import json
-import os
-import shutil
-import threading
 from concurrent.futures import ThreadPoolExecutor
 from pathlib import Path
-from unittest.mock import MagicMock, patch
+from unittest.mock import patch
 
-import pytest
-
-from scripts.run_eval import (
-    run_eval,
-    run_single_query,
-    run_single_query_claude,
-    run_single_query_codex,
-)
+from scripts.run_eval import run_eval, run_single_query
 from scripts.utils import CLI_CLAUDE, CLI_CODEX
-
-SKILL_NAME = "skill-forge"
 
 
 class TestRunSingleQueryDispatch:
@@ -55,571 +42,6 @@ class TestRunSingleQueryDispatch:
             )
             args, _ = mock.call_args
             assert args[-1] == "/custom/codex"
-
-
-class TestCodexCommandArgs:
-    """Verify that the codex exec command is built with valid arguments."""
-
-    def test_no_invalid_approval_flag(self, tmp_path):
-        """Ensure -a flag is not used (removed in current codex CLI)."""
-        project_root = tmp_path / "project"
-        (project_root / ".agents" / "skills").mkdir(parents=True)
-
-        captured_cmd = []
-
-        def capture_popen(cmd, **kwargs):
-            captured_cmd.extend(cmd)
-            mock_proc = MagicMock()
-            mock_proc.poll.side_effect = [0, 0]
-            mock_proc.stdout.read.return_value = b'{"type":"turn.completed"}\n'
-            mock_proc.stderr.read.return_value = b""
-            mock_proc.returncode = 0
-            return mock_proc
-
-        with patch("scripts.run_eval_codex.subprocess.Popen", side_effect=capture_popen):
-            with patch("scripts.trigger_probe.select.select", return_value=([], [], [])):
-                run_single_query_codex(
-                    "test query", "my-skill", "test desc", 5, str(project_root),
-                )
-
-        assert "-a" not in captured_cmd, "codex exec should not use -a flag"
-        assert "never" not in captured_cmd, "codex exec should not use 'never' argument"
-
-    def test_codex_command_includes_required_flags(self, tmp_path):
-        """Verify codex exec includes --json, -s, -C flags."""
-        project_root = tmp_path / "project"
-        (project_root / ".agents" / "skills").mkdir(parents=True)
-
-        captured_cmd = []
-
-        def capture_popen(cmd, **kwargs):
-            captured_cmd.extend(cmd)
-            mock_proc = MagicMock()
-            mock_proc.poll.side_effect = [0, 0]
-            mock_proc.stdout.read.return_value = b'{"type":"turn.completed"}\n'
-            mock_proc.stderr.read.return_value = b""
-            mock_proc.returncode = 0
-            return mock_proc
-
-        with patch("scripts.run_eval_codex.subprocess.Popen", side_effect=capture_popen):
-            with patch("scripts.trigger_probe.select.select", return_value=([], [], [])):
-                run_single_query_codex(
-                    "test query", "my-skill", "test desc", 5, str(project_root),
-                )
-
-        assert "exec" in captured_cmd
-        assert "--json" in captured_cmd
-        assert "-s" in captured_cmd
-        assert "read-only" in captured_cmd
-        assert "-C" in captured_cmd
-
-
-class TestCliExitCodeHandling:
-    """Verify that non-zero CLI exit codes raise RuntimeError."""
-
-    def test_codex_nonzero_exit_raises(self, tmp_path):
-        project_root = tmp_path / "project"
-        (project_root / ".agents" / "skills").mkdir(parents=True)
-
-        mock_process = MagicMock()
-        mock_process.poll.side_effect = [0, 0]
-        mock_process.stdout.read.return_value = b""
-        mock_process.stdout.fileno.return_value = 0
-        mock_process.stderr.read.return_value = b"codex: unknown option '-a'"
-        mock_process.returncode = 1
-
-        with patch("scripts.run_eval_codex.subprocess.Popen", return_value=mock_process):
-            with patch("scripts.trigger_probe.select.select", return_value=([], [], [])):
-                with pytest.raises(RuntimeError, match="Codex CLI exited with code 1"):
-                    run_single_query_codex(
-                        "test query", "my-skill", "test desc", 5, str(project_root),
-                    )
-
-    def test_run_eval_tracks_errors_in_summary(self):
-        """Verify that CLI errors are counted in summary.errors."""
-        eval_set = [
-            {"query": "error query", "should_trigger": True},
-            {"query": "ok query", "should_trigger": True},
-        ]
-
-        def mock_run(query, *args, **kwargs):
-            if query == "error query":
-                raise RuntimeError("CLI crashed")
-            return "triggered"
-
-        with patch("scripts.run_eval.ProcessPoolExecutor", ThreadPoolExecutor):
-            with patch("scripts.run_eval.run_single_query", side_effect=mock_run):
-                result = run_eval(
-                    eval_set=eval_set,
-                    skill_name="test",
-                    description="test desc",
-                    num_workers=1,
-                    timeout=10,
-                    project_root=Path("/tmp"),
-                    runs_per_query=1,
-                    cli_type=CLI_CLAUDE,
-                )
-
-        assert result["summary"]["errors"] == 1
-        error_results = [r for r in result["results"] if r.get("errors")]
-        assert len(error_results) == 1
-        assert "CLI crashed" in error_results[0]["errors"][0]
-
-
-class TestRunSingleQueryClaude:
-    def _make_process_mock(self, output_lines: list[str]):
-        output = ("\n".join(output_lines) + "\n").encode()
-        mock_process = MagicMock()
-        mock_process.poll.side_effect = [None, 0]
-        mock_process.stdout.read.return_value = b""
-        mock_process.stdout.fileno.return_value = 0
-        mock_process.stderr.read.return_value = b""
-        mock_process.returncode = 0
-        return mock_process, output
-
-    def test_detects_canonical_skill_name_from_assistant_tool_use(self, tmp_path):
-        project_root = tmp_path / "project"
-        project_root.mkdir()
-
-        events = [
-            json.dumps({
-                "type": "assistant",
-                "message": {
-                    "content": [
-                        {"type": "tool_use", "name": "Skill", "input": {"skill": SKILL_NAME}},
-                    ],
-                },
-            }),
-            json.dumps({"type": "result"}),
-        ]
-        mock_process, output = self._make_process_mock(events)
-
-        with patch("scripts.run_eval_claude.subprocess.Popen", return_value=mock_process):
-            with patch("scripts.trigger_probe.select.select", return_value=([mock_process.stdout], [], [])):
-                with patch("scripts.trigger_probe.os.read", return_value=output):
-                    result = run_single_query_claude(
-                        "test query", SKILL_NAME, "test desc", 5, str(project_root),
-                    )
-
-        assert result == "triggered"
-
-    def test_detects_stream_event_skill_tool_use(self, tmp_path):
-        project_root = tmp_path / "project"
-        project_root.mkdir()
-
-        events = [
-            json.dumps({
-                "type": "stream_event",
-                "event": {
-                    "type": "content_block_start",
-                    "content_block": {"type": "tool_use", "name": "Skill"},
-                },
-            }),
-            json.dumps({
-                "type": "stream_event",
-                "event": {
-                    "type": "content_block_delta",
-                    "delta": {
-                        "type": "input_json_delta",
-                        "partial_json": json.dumps({"skill": SKILL_NAME}),
-                    },
-                },
-            }),
-            json.dumps({
-                "type": "stream_event",
-                "event": {"type": "content_block_stop"},
-            }),
-        ]
-        mock_process, output = self._make_process_mock(events)
-
-        with patch("scripts.run_eval_claude.subprocess.Popen", return_value=mock_process):
-            with patch("scripts.trigger_probe.select.select", return_value=([mock_process.stdout], [], [])):
-                with patch("scripts.trigger_probe.os.read", return_value=output):
-                    result = run_single_query_claude(
-                        "test query", SKILL_NAME, "test desc", 5, str(project_root),
-                    )
-
-        assert result == "triggered"
-
-    def test_stream_event_shape_change_does_not_trigger(self, tmp_path):
-        project_root = tmp_path / "project"
-        project_root.mkdir()
-
-        events = [
-            json.dumps({
-                "type": "stream_event",
-                "event": {
-                    "type": "content_block_delta",
-                    "delta": {
-                        "type": "text_delta",
-                        "text": json.dumps({"skill": SKILL_NAME}),
-                    },
-                },
-            }),
-            json.dumps({"type": "result"}),
-        ]
-        mock_process, output = self._make_process_mock(events)
-
-        with patch("scripts.run_eval_claude.subprocess.Popen", return_value=mock_process):
-            with patch("scripts.trigger_probe.select.select", return_value=([mock_process.stdout], [], [])):
-                with patch("scripts.trigger_probe.os.read", return_value=output):
-                    result = run_single_query_claude(
-                        "test query", SKILL_NAME, "test desc", 5, str(project_root),
-                    )
-
-        assert result == "completed"
-
-    def test_ignores_non_matching_tool_before_skill_trigger(self, tmp_path):
-        project_root = tmp_path / "project"
-        project_root.mkdir()
-
-        events = [
-            json.dumps({
-                "type": "assistant",
-                "message": {
-                    "content": [
-                        {"type": "tool_use", "name": "Bash", "input": {"command": "pwd"}},
-                    ],
-                },
-            }),
-            json.dumps({
-                "type": "assistant",
-                "message": {
-                    "content": [
-                        {"type": "tool_use", "name": "Skill", "input": {"skill": SKILL_NAME}},
-                    ],
-                },
-            }),
-            json.dumps({"type": "result"}),
-        ]
-        mock_process, output = self._make_process_mock(events)
-
-        with patch("scripts.run_eval_claude.subprocess.Popen", return_value=mock_process):
-            with patch("scripts.trigger_probe.select.select", return_value=([mock_process.stdout], [], [])):
-                with patch("scripts.trigger_probe.os.read", return_value=output):
-                    result = run_single_query_claude(
-                        "test query", SKILL_NAME, "test desc", 5, str(project_root),
-                    )
-
-        assert result == "triggered"
-
-    def test_detects_skill_path_from_read_tool_use(self, tmp_path):
-        project_root = tmp_path / "project"
-        project_root.mkdir()
-
-        events = [
-            json.dumps({
-                "type": "assistant",
-                "message": {
-                    "content": [
-                        {
-                            "type": "tool_use",
-                            "name": "Read",
-                            "input": {
-                                "file_path": str(project_root / ".claude" / "skills" / SKILL_NAME / "SKILL.md"),
-                            },
-                        },
-                    ],
-                },
-            }),
-            json.dumps({"type": "result"}),
-        ]
-        mock_process, output = self._make_process_mock(events)
-
-        with patch("scripts.run_eval_claude.subprocess.Popen", return_value=mock_process):
-            with patch("scripts.trigger_probe.select.select", return_value=([mock_process.stdout], [], [])):
-                with patch("scripts.trigger_probe.os.read", return_value=output):
-                    result = run_single_query_claude(
-                        "test query", SKILL_NAME, "test desc", 5, str(project_root),
-                    )
-
-        assert result == "triggered"
-
-    def test_uses_isolated_workspace_for_temp_skill_even_with_override(self, tmp_path):
-        project_root = tmp_path / "project"
-        project_root.mkdir()
-        claude_home = tmp_path / "custom-claude-home"
-        observed = {}
-
-        events = [
-            json.dumps({"type": "result"}),
-        ]
-        mock_process, output = self._make_process_mock(events)
-
-        with patch.dict(os.environ, {"SKILL_FORGE_CLAUDE_HOME": str(claude_home)}, clear=True):
-            def capture_popen(*args, **kwargs):
-                observed["cwd"] = Path(kwargs["cwd"])
-                observed["claude_home"] = Path(kwargs["env"]["SKILL_FORGE_CLAUDE_HOME"])
-                observed["claude_config_dir"] = Path(kwargs["env"]["CLAUDE_CONFIG_DIR"])
-                return mock_process
-
-            with patch("scripts.run_eval_claude.subprocess.Popen", side_effect=capture_popen):
-                with patch("scripts.trigger_probe.select.select", return_value=([mock_process.stdout], [], [])):
-                    with patch("scripts.trigger_probe.os.read", return_value=output):
-                        result = run_single_query_claude(
-                            "test query", SKILL_NAME, "test desc", 5, str(project_root),
-                        )
-
-        assert result == "completed"
-        assert observed["cwd"] == project_root
-        assert observed["claude_home"].parent == project_root
-        assert observed["claude_config_dir"] == observed["claude_home"]
-        assert not observed["claude_home"].exists()
-        assert not (claude_home / "commands").exists()
-        assert not (claude_home / "skills").exists()
-        assert not (project_root / ".claude").exists()
-
-    def test_copies_supporting_files_into_temp_skill(self, tmp_path):
-        project_root = tmp_path / "project"
-        source_skill = project_root / ".claude" / "skills" / SKILL_NAME
-        source_skill.mkdir(parents=True)
-        (source_skill / "SKILL.md").write_text(
-            "---\n"
-            f"name: {SKILL_NAME}\n"
-            "description: old desc\n"
-            "---\n\n"
-            "# Old\n"
-        )
-        (source_skill / "references").mkdir()
-        (source_skill / "references" / "guide.md").write_text("supporting file")
-
-        observed = {}
-
-        events = [
-            json.dumps({"type": "result"}),
-        ]
-        mock_process, output = self._make_process_mock(events)
-
-        def capture_popen(*args, **kwargs):
-            temp_skill = Path(kwargs["env"]["SKILL_FORGE_CLAUDE_HOME"]) / "skills" / SKILL_NAME
-            observed["skill_md"] = (temp_skill / "SKILL.md").read_text()
-            observed["supporting_file"] = (temp_skill / "references" / "guide.md").read_text()
-            return mock_process
-
-        with patch("scripts.run_eval_claude.subprocess.Popen", side_effect=capture_popen):
-            with patch("scripts.trigger_probe.select.select", return_value=([mock_process.stdout], [], [])):
-                with patch("scripts.trigger_probe.os.read", return_value=output):
-                    result = run_single_query_claude(
-                        "test query", SKILL_NAME, "test desc", 5, str(project_root),
-                    )
-
-        assert result == "completed"
-        assert "description: |\n  test desc\n" in observed["skill_md"]
-        assert "old desc" not in observed["skill_md"]
-        assert observed["supporting_file"] == "supporting file"
-
-    def test_detects_relative_skill_path_from_read_tool_use(self, tmp_path):
-        project_root = tmp_path / "project"
-        project_root.mkdir()
-        claude_home = tmp_path / "custom-claude-home"
-
-        events = [
-            json.dumps({
-                "type": "assistant",
-                "message": {
-                    "content": [
-                        {
-                            "type": "tool_use",
-                            "name": "Read",
-                            "input": {
-                                "file_path": f"skills/{SKILL_NAME}/SKILL.md",
-                            },
-                        },
-                    ],
-                },
-            }),
-            json.dumps({"type": "result"}),
-        ]
-        mock_process, output = self._make_process_mock(events)
-
-        with patch.dict(os.environ, {"SKILL_FORGE_CLAUDE_HOME": str(claude_home)}, clear=True):
-            with patch("scripts.run_eval_claude.subprocess.Popen", return_value=mock_process):
-                with patch("scripts.trigger_probe.select.select", return_value=([mock_process.stdout], [], [])):
-                    with patch("scripts.trigger_probe.os.read", return_value=output):
-                        result = run_single_query_claude(
-                            "test query", SKILL_NAME, "test desc", 5, str(project_root),
-                        )
-
-        assert result == "triggered"
-
-
-class TestRunSingleQueryCodex:
-    def _make_process_mock(self, output_lines: list[str]):
-        """Create a mock process that yields output_lines then exits."""
-        output = ("\n".join(output_lines) + "\n").encode()
-        mock_process = MagicMock()
-        # First poll returns None (running), then 0 (done), then 0 (finally block check)
-        mock_process.poll.side_effect = [None, 0, 0]
-        mock_process.stdout.read.return_value = output
-        mock_process.stdout.fileno.return_value = 0
-        mock_process.stderr.read.return_value = b""
-        mock_process.returncode = 0
-        return mock_process
-
-    def test_creates_and_cleans_temp_skill(self, tmp_path):
-        """Verify temp skill dir is created and cleaned up."""
-        project_root = tmp_path / "project"
-        (project_root / ".agents" / "skills").mkdir(parents=True)
-
-        events = [
-            json.dumps({"type": "turn.completed", "usage": {}}),
-        ]
-        mock_process = self._make_process_mock(events)
-
-        with patch("scripts.run_eval_codex.subprocess.Popen", return_value=mock_process):
-            with patch("scripts.trigger_probe.select.select", return_value=([], [], [])):
-                result = run_single_query_codex(
-                    "test query", "my-skill", "test desc", 5, str(project_root),
-                )
-
-        assert result == "completed"
-        # Temp skill dir should be cleaned up
-        skill_dirs = list((project_root / ".agents" / "skills").iterdir())
-        assert len(skill_dirs) == 0
-
-    def test_creates_temp_skill_in_repo_agents_even_with_codex_home_override(self, tmp_path):
-        project_root = tmp_path / "project"
-        project_root.mkdir(parents=True)
-        codex_home = tmp_path / "custom-codex-home"
-
-        events = [
-            json.dumps({"type": "turn.completed", "usage": {}}),
-        ]
-        mock_process = self._make_process_mock(events)
-
-        with patch.dict(os.environ, {"CODEX_HOME": str(codex_home)}, clear=True):
-            with patch("scripts.run_eval_codex.subprocess.Popen", return_value=mock_process):
-                with patch("scripts.trigger_probe.select.select", return_value=([], [], [])):
-                    result = run_single_query_codex(
-                        "test query", "my-skill", "test desc", 5, str(project_root),
-                    )
-
-        assert result == "completed"
-        assert not (codex_home / "skills").exists()
-        assert not (project_root / ".codex").exists()
-        assert (project_root / ".agents" / "skills").is_dir()
-        assert not any((project_root / ".agents" / "skills").iterdir())
-
-    def test_detects_marker_in_agent_message(self, tmp_path):
-        """Verify marker detection in codex JSONL output."""
-        project_root = tmp_path / "project"
-        (project_root / ".agents" / "skills").mkdir(parents=True)
-
-        with patch("scripts.run_eval_codex.uuid.uuid4") as mock_uuid:
-            mock_uuid.return_value.hex = "abcd1234xxxxxxxxxxxxxxxx"
-            marker = "[SKILL_TRIGGERED:abcd1234]"
-
-            events = [
-                json.dumps({"type": "item.completed", "item": {"type": "agent_message", "text": f"Result {marker} done."}}),
-                json.dumps({"type": "turn.completed", "usage": {}}),
-            ]
-            output = ("\n".join(events) + "\n").encode()
-
-            mock_process = MagicMock()
-            # Return None first (process running), read stdout, then 0 (done)
-            mock_process.poll.side_effect = [None, 0]
-            mock_process.stdout.fileno.return_value = 0
-            mock_process.stdout.read.return_value = output
-            mock_process.stderr.read.return_value = b""
-            mock_process.returncode = 0
-
-            with patch("scripts.run_eval_codex.subprocess.Popen", return_value=mock_process):
-                with patch("scripts.trigger_probe.select.select", return_value=([mock_process.stdout], [], [])):
-                    with patch("scripts.trigger_probe.os.read", return_value=output):
-                        result = run_single_query_codex(
-                            "test query", "my-skill", "test desc", 5, str(project_root),
-                        )
-
-            assert result == "triggered"
-
-    def test_detects_marker_in_updated_agent_message(self, tmp_path):
-        project_root = tmp_path / "project"
-        (project_root / ".agents" / "skills").mkdir(parents=True)
-
-        with patch("scripts.run_eval_codex.uuid.uuid4") as mock_uuid:
-            mock_uuid.return_value.hex = "abcd1234xxxxxxxxxxxxxxxx"
-            marker = "[SKILL_TRIGGERED:abcd1234]"
-
-            events = [
-                json.dumps({"type": "item.updated", "item": {"type": "agent_message", "text": f"{marker}"}}),
-            ]
-            output = ("\n".join(events) + "\n").encode()
-
-            mock_process = MagicMock()
-            mock_process.poll.side_effect = [None, 0]
-            mock_process.stdout.fileno.return_value = 0
-            mock_process.stdout.read.return_value = output
-            mock_process.stderr.read.return_value = b""
-            mock_process.returncode = 0
-
-            with patch("scripts.run_eval_codex.subprocess.Popen", return_value=mock_process):
-                with patch("scripts.trigger_probe.select.select", return_value=([mock_process.stdout], [], [])):
-                    with patch("scripts.trigger_probe.os.read", return_value=output):
-                        result = run_single_query_codex(
-                            "test query", "my-skill", "test desc", 5, str(project_root),
-                        )
-
-            assert result == "triggered"
-
-    def test_codex_event_shape_change_does_not_trigger(self, tmp_path):
-        project_root = tmp_path / "project"
-        (project_root / ".agents" / "skills").mkdir(parents=True)
-
-        with patch("scripts.run_eval_codex.uuid.uuid4") as mock_uuid:
-            mock_uuid.return_value.hex = "abcd1234xxxxxxxxxxxxxxxx"
-            marker = "[SKILL_TRIGGERED:abcd1234]"
-
-            events = [
-                json.dumps({
-                    "type": "message.completed",
-                    "message": {"role": "assistant", "content": f"{marker}"},
-                }),
-                json.dumps({"type": "turn.completed", "usage": {}}),
-            ]
-            output = ("\n".join(events) + "\n").encode()
-
-            mock_process = MagicMock()
-            mock_process.poll.side_effect = [None, 0]
-            mock_process.stdout.fileno.return_value = 0
-            mock_process.stdout.read.return_value = output
-            mock_process.stderr.read.return_value = b""
-            mock_process.returncode = 0
-
-            with patch("scripts.run_eval_codex.subprocess.Popen", return_value=mock_process):
-                with patch("scripts.trigger_probe.select.select", return_value=([mock_process.stdout], [], [])):
-                    with patch("scripts.trigger_probe.os.read", return_value=output):
-                        result = run_single_query_codex(
-                            "test query", "my-skill", "test desc", 5, str(project_root),
-                        )
-
-            assert result == "completed"
-
-    def test_no_trigger_returns_false(self, tmp_path):
-        """No marker in output means not triggered."""
-        project_root = tmp_path / "project"
-        (project_root / ".agents" / "skills").mkdir(parents=True)
-
-        events = [
-            json.dumps({"type": "item.completed", "item": {"type": "agent_message", "text": "No skill here."}}),
-            json.dumps({"type": "turn.completed", "usage": {}}),
-        ]
-        output = ("\n".join(events) + "\n").encode()
-
-        mock_process = MagicMock()
-        mock_process.poll.side_effect = [None, 0]
-        mock_process.stdout.fileno.return_value = 0
-        mock_process.stdout.read.return_value = output
-        mock_process.stderr.read.return_value = b""
-        mock_process.returncode = 0
-
-        with patch("scripts.run_eval_codex.subprocess.Popen", return_value=mock_process):
-            with patch("scripts.trigger_probe.select.select", return_value=([mock_process.stdout], [], [])):
-                with patch("scripts.trigger_probe.os.read", return_value=output):
-                    result = run_single_query_codex(
-                        "test query", "my-skill", "test desc", 5, str(project_root),
-                    )
-
-        assert result == "completed"
 
 
 class TestRunEval:
@@ -808,183 +230,35 @@ class TestRunEval:
         assert result["results"][0]["status"] == "not_run"
         assert result["results"][0]["pass"] is False
 
+    def test_run_eval_tracks_errors_in_summary(self):
+        """Verify that CLI errors are counted in summary.errors."""
+        eval_set = [
+            {"query": "error query", "should_trigger": True},
+            {"query": "ok query", "should_trigger": True},
+        ]
 
-class TestTemporarySkillNames:
-    def test_codex_temp_skill_keeps_original_visible_name(self, tmp_path):
-        project_root = tmp_path / "project"
-        (project_root / ".agents" / "skills").mkdir(parents=True)
+        def mock_run(query, *args, **kwargs):
+            if query == "error query":
+                raise RuntimeError("CLI crashed")
+            return "triggered"
 
-        observed = {}
-
-        def capture_popen(cmd, **kwargs):
-            skill_root = project_root / ".agents" / "skills"
-            skill_file = next(skill_root.glob("*/SKILL.md"))
-            observed["path"] = skill_file
-            observed["content"] = skill_file.read_text()
-
-            mock_process = MagicMock()
-            mock_process.poll.side_effect = [0, 0]
-            mock_process.stdout.read.return_value = b'{"type":"turn.completed"}\n'
-            mock_process.stderr.read.return_value = b""
-            mock_process.returncode = 0
-            return mock_process
-
-        with patch.dict(os.environ, {}, clear=True):
-            with patch("scripts.run_eval_codex.subprocess.Popen", side_effect=capture_popen):
-                with patch("scripts.trigger_probe.select.select", return_value=([], [], [])):
-                    run_single_query_codex(
-                        "test query", SKILL_NAME, "test desc", 5, str(project_root),
-                    )
-
-        assert observed["path"].parent.name.startswith(f"{SKILL_NAME}-skill-")
-        assert "name: skill-forge\n" in observed["content"]
-        assert f"name: {SKILL_NAME}-skill-" not in observed["content"]
-
-    def test_claude_temp_skill_keeps_original_visible_name(self, tmp_path):
-        project_root = tmp_path / "project"
-        project_root.mkdir()
-
-        observed = {}
-
-        def capture_popen(cmd, **kwargs):
-            skill_root = Path(kwargs["env"]["SKILL_FORGE_CLAUDE_HOME"]) / "skills"
-            skill_file = next(skill_root.glob("*/SKILL.md"))
-            observed["path"] = skill_file
-            observed["content"] = skill_file.read_text()
-
-            mock_process = MagicMock()
-            mock_process.poll.side_effect = [0, 0]
-            mock_process.stdout.read.return_value = b'{"type":"result"}\n'
-            mock_process.stderr.read.return_value = b""
-            mock_process.returncode = 0
-            return mock_process
-
-        with patch("scripts.run_eval_claude.subprocess.Popen", side_effect=capture_popen):
-            with patch("scripts.trigger_probe.select.select", return_value=([], [], [])):
-                run_single_query_claude(
-                    "test query", SKILL_NAME, "test desc", 5, str(project_root),
+        with patch("scripts.run_eval.ProcessPoolExecutor", ThreadPoolExecutor):
+            with patch("scripts.run_eval.run_single_query", side_effect=mock_run):
+                result = run_eval(
+                    eval_set=eval_set,
+                    skill_name="test",
+                    description="test desc",
+                    num_workers=1,
+                    timeout=10,
+                    project_root=Path("/tmp"),
+                    runs_per_query=1,
+                    cli_type=CLI_CLAUDE,
                 )
 
-        assert observed["path"].parent.name == SKILL_NAME
-        assert observed["path"].name == "SKILL.md"
-        assert f"name: {SKILL_NAME}\n" in observed["content"]
-        assert f"# {SKILL_NAME}\n" in observed["content"]
-
-    def test_codex_temp_skill_paths_are_isolated_per_run(self, tmp_path):
-        project_root = tmp_path / "project"
-        (project_root / ".agents" / "skills").mkdir(parents=True)
-
-        observed_snapshots = []
-        observed_lock = threading.Lock()
-        barrier = threading.Barrier(2)
-
-        def capture_popen(cmd, **kwargs):
-            skill_root = project_root / ".agents" / "skills"
-            barrier.wait(timeout=2)
-            snapshot = sorted(path.name for path in skill_root.iterdir())
-            with observed_lock:
-                observed_snapshots.append(snapshot)
-
-            mock_process = MagicMock()
-            mock_process.poll.side_effect = [0, 0]
-            mock_process.stdout.read.return_value = b'{"type":"turn.completed"}\n'
-            mock_process.stderr.read.return_value = b""
-            mock_process.returncode = 0
-            return mock_process
-
-        with patch.dict(os.environ, {}, clear=True):
-            with patch("scripts.run_eval_codex.subprocess.Popen", side_effect=capture_popen):
-                with patch("scripts.trigger_probe.select.select", return_value=([], [], [])):
-                    with ThreadPoolExecutor(max_workers=2) as executor:
-                        futures = [
-                            executor.submit(
-                                run_single_query_codex,
-                                "test query",
-                                SKILL_NAME,
-                                "test desc",
-                                5,
-                                str(project_root),
-                            )
-                            for _ in range(2)
-                        ]
-                        for future in futures:
-                            assert future.result() == "completed"
-
-        assert len(observed_snapshots) == 2
-        assert all(len(snapshot) == 2 for snapshot in observed_snapshots)
-        assert all(len(set(snapshot)) == 2 for snapshot in observed_snapshots)
-
-    def test_claude_temp_skill_paths_are_isolated_per_run(self, tmp_path):
-        project_root = tmp_path / "project"
-        project_root.mkdir()
-
-        observed_paths = []
-        observed_lock = threading.Lock()
-        barrier = threading.Barrier(2)
-
-        def capture_popen(cmd, **kwargs):
-            claude_home = Path(kwargs["env"]["SKILL_FORGE_CLAUDE_HOME"])
-            skill_root = claude_home / "skills"
-            barrier.wait(timeout=2)
-            snapshot = sorted(path.name for path in skill_root.iterdir())
-            with observed_lock:
-                observed_paths.append((claude_home.name, snapshot))
-
-            mock_process = MagicMock()
-            mock_process.poll.side_effect = [0, 0]
-            mock_process.stdout.read.return_value = b'{"type":"result"}\n'
-            mock_process.stderr.read.return_value = b""
-            mock_process.returncode = 0
-            return mock_process
-
-        with patch("scripts.run_eval_claude.subprocess.Popen", side_effect=capture_popen):
-            with patch("scripts.trigger_probe.select.select", return_value=([], [], [])):
-                with ThreadPoolExecutor(max_workers=2) as executor:
-                    futures = [
-                        executor.submit(
-                            run_single_query_claude,
-                            "test query",
-                            SKILL_NAME,
-                            "test desc",
-                            5,
-                            str(project_root),
-                        )
-                        for _ in range(2)
-                    ]
-                    for future in futures:
-                        assert future.result() == "completed"
-
-        assert len(observed_paths) == 2
-        assert len({path for path, _ in observed_paths}) == 2
-        assert all(snapshot == [SKILL_NAME] for _, snapshot in observed_paths)
-
-    def test_claude_runs_in_project_root_with_isolated_home(self, tmp_path):
-        project_root = tmp_path / "project"
-        project_root.mkdir()
-
-        observed = {}
-
-        def capture_popen(cmd, **kwargs):
-            observed["cwd"] = Path(kwargs["cwd"])
-            observed["claude_home"] = Path(kwargs["env"]["SKILL_FORGE_CLAUDE_HOME"])
-
-            mock_process = MagicMock()
-            mock_process.poll.side_effect = [0, 0]
-            mock_process.stdout.read.return_value = b'{"type":"result"}\n'
-            mock_process.stderr.read.return_value = b""
-            mock_process.returncode = 0
-            return mock_process
-
-        with patch("scripts.run_eval_claude.subprocess.Popen", side_effect=capture_popen):
-            with patch("scripts.trigger_probe.select.select", return_value=([], [], [])):
-                result = run_single_query_claude(
-                    "test query", SKILL_NAME, "test desc", 5, str(project_root),
-                )
-
-        assert result == "completed"
-        assert observed["cwd"] == project_root
-        assert observed["claude_home"].parent == project_root
-        assert not observed["claude_home"].exists()
+        assert result["summary"]["errors"] == 1
+        error_results = [r for r in result["results"] if r.get("errors")]
+        assert len(error_results) == 1
+        assert "CLI crashed" in error_results[0]["errors"][0]
 
 
 class TestCodexSerialization:
@@ -1037,37 +311,7 @@ class TestCodexSerialization:
         assert captured["max_workers"] == 8
 
 
-class TestTimeoutOutcome:
-    def test_claude_timeout_returns_timeout_outcome(self, tmp_path):
-        project_root = tmp_path / "project"
-        project_root.mkdir()
-
-        mock_process = MagicMock()
-        mock_process.poll.return_value = None
-
-        with patch("scripts.run_eval_claude.subprocess.Popen", return_value=mock_process):
-            result = run_single_query_claude(
-                "test query", SKILL_NAME, "test desc", 0, str(project_root),
-            )
-
-        assert result == "timeout"
-        mock_process.kill.assert_called_once()
-
-    def test_codex_timeout_returns_timeout_outcome(self, tmp_path):
-        project_root = tmp_path / "project"
-        (project_root / ".agents" / "skills").mkdir(parents=True)
-
-        mock_process = MagicMock()
-        mock_process.poll.return_value = None
-
-        with patch("scripts.run_eval_codex.subprocess.Popen", return_value=mock_process):
-            result = run_single_query_codex(
-                "test query", "my-skill", "test desc", 0, str(project_root),
-            )
-
-        assert result == "timeout"
-        mock_process.kill.assert_called_once()
-
+class TestTimeoutAggregation:
     def test_run_eval_counts_timeouts_as_non_triggers(self):
         eval_set = [{"query": "slow query", "should_trigger": True}]
